Cleanup
Renamed class to Main
Moved a method to Settings class
